requirements.txt
psycopg2-binary
django
djangorestframework
markdown
django-cors-headers
django-webpack-loader
uvicorn
python-dateutil
numpy
rasterio
coverage
pyspark